Create method to convert date to timestamp
This method will be utilized in the Ticker class, specifically the history method.  It allows the user to specify a start and end date in either YYYY-MM-DD format or as a datetime object.  Additionally, it will supply default values in the event the period argument equals max (defaults to 1900-01-01 for start and now for end).